properties/views.py
from django.shortcuts import render

# Create your views here.
from rest_framework import viewsets, filters
from django_filters.rest_framework import DjangoFilterBackend
from .models import Property
from .serializers import PropertySerializer

class PropertyViewSet(viewsets.ReadOnlyModelViewSet):
    queryset = Property.objects.all()
    serializer_class = PropertySerializer
    filter_backends = [DjangoFilterBackend, filters.SearchFilter, filters.OrderingFilter]
    filterset_fields = ['city', 'zip_code', 'transaction_type', 'property_type', 'status']
    search_fields = ['title', 'description', 'city']
    ordering_fields = ['price', 'created_at', 'surface']
    ordering = ['-created_at']